Added Recipe content and Links
Index.html was a blank project with nothing in it. This was the starting place.

Added in a header to index.html, as well as a link to the the Pizza Recipe page.
The pizza recipe page was created and a header was added to tell
the reader this is the pizza recipe page.

This was done according to the project guidelines.